Move pickle migration script to its own .py script.
